// src/models/supplier.tsx
class Supplier {
    "Name": string;
    "Email": string;
    "Phone": string; 
    "Cellphone": string; 
    "ID": number | null;
    "CompanyID": number | null;
    "CompanyName": string;
    "PersonType": string;
    "Street": string;
    "Number": number | null;
    "Neighborhood": string;
    "State": string;
    "City": string;
    "Address": string;
    "CEP": number | null;
}

export default Supplier;